parking_agent/pyproject.toml
[project]
name = "parking_agent"
version = "0.1.0"
description = "parking-agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.186.1,<1.0.0",
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pytz>=2023.3",
    "pydantic>=2.5.0"
]

[project.scripts]
parking_agent = "parking_agent.main:run"
run_crew = "parking_agent.main:run"
train = "parking_agent.main:train"
replay = "parking_agent.main:replay"
test = "parking_agent.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
